added definitions file
